fixed useContext/invalid hooks bug
