+++ components/ClientLayout.tsx
@@ -1,0 +1,39 @@
+"use client";
+
+import { useEffect } from "react";
+import { userAtom } from "@/atoms/user/atom";
+import { useAtom } from "jotai";
+import { get } from "@/app/lib/fetchInterceptor";
+import { useTranslation } from "next-i18next";
+
+export default function ClientLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const { i18n } = useTranslation();
+  const [, setUserData] = useAtom(userAtom);
+
+  useEffect(() => {
+    // read the token only once on mount
+    const token = localStorage.getItem("accessToken");
+    if (!token) return;
+
+    const fetchData = async () => {
+      try {
+        const data = await get("/api/member/me");
+        setUserData(data);
+      } catch (error) {
+        console.error("Failed to fetch profile data:", error);
+      }
+    };
+
+    fetchData();
+  }, [i18n.language, setUserData]);
+
+  return (
+    <div className="antialiased min-h-screen flex flex-col p-12">
+      <main className="flex-grow bg-white">{children}</main>
+    </div>
+  );
+}
